fix(acip): change STREAM_START/STOP to uint32_t
Protocol fix: STREAM_START and STREAM_STOP handlers expected uint8_t
but acip_send_stream_start() in send.c sends uint32_t (4 bytes),
causing "payload size 1 (expected 4)" errors.

Root cause: Type mismatch across protocol layers:
- Client (send.c): Sends 4-byte uint32_t in network byte order
- Server (handlers.c): Expected 1-byte uint8_t
- Result: Size validation failure, packet rejected

Solution: Update handlers to match the send implementation:
1. Change payload size check from sizeof(uint8_t) to sizeof(uint32_t)
2. Read 4 bytes and convert from network byte order
3. Update callback signatures to accept uint32_t

Changes:
- handlers.h: Update on_stream_start/stop signatures (uint8_t → uint32_t)
- handlers.c: Read uint32_t with network byte order conversion
- handlers.c: Add better error messages with actual/expected sizes

Impact: Fixes STREAM_START/STOP protocol compatibility

+++ lib/network/acip/handlers.h
@@ -152,10 +152,10 @@ typedef struct {
   void (*on_client_leave)(void *client_ctx, void *app_ctx);
 
   /** @brief Called when client starts streaming */
-  void (*on_stream_start)(uint8_t stream_types, void *client_ctx, void *app_ctx);
+  void (*on_stream_start)(uint32_t stream_types, void *client_ctx, void *app_ctx);
 
   /** @brief Called when client stops streaming */
-  void (*on_stream_stop)(uint8_t stream_types, void *client_ctx, void *app_ctx);
+  void (*on_stream_stop)(uint32_t stream_types, void *client_ctx, void *app_ctx);
 
   /** @brief Called when client sends capabilities */
   void (*on_capabilities)(const void *cap_data, size_t data_len, void *client_ctx, void *app_ctx);
